subdomain-finder/requirements.txt
flask
requests
python-whois
